removed references to avatar
