+++ GRA.Database/dbo/Stored Procedures/app_AvatarPart_Delete.sql
@@ -1,0 +1,8 @@
+﻿
+CREATE PROCEDURE [dbo].[app_AvatarPart_Delete] @APID INT,
+	@TenID INT = NULL
+AS
+DELETE
+FROM [AvatarPart]
+WHERE APID = @APID
+	AND TenID = @TenID
