+++ GRA.Database/dbo/Stored Procedures/app_AvatarPart_Insert.sql
