--- GRA.Database/dbo/Stored Procedures/app_Avatar_Delete.sql
@@ -1,8 +1,0 @@
-﻿
-CREATE PROCEDURE [dbo].[app_Avatar_Delete] @AID INT,
-	@TenID INT = NULL
-AS
-DELETE
-FROM [Avatar]
-WHERE AID = @AID
-	AND TenID = @TenID
