--- GRA.Database/dbo/Stored Procedures/app_Avatar_GetAll.sql
@@ -1,9 +1,0 @@
-﻿
-CREATE PROCEDURE [dbo].[app_Avatar_GetAll] @TenID INT = NULL
-AS
-SELECT *
-FROM [Avatar]
-WHERE (
-		TenID = @TenID
-		OR @TenID IS NULL
-		)
