--- GRA.Database/dbo/Stored Procedures/app_Avatar_GetByID.sql
@@ -1,6 +1,0 @@
-﻿
-CREATE PROCEDURE [dbo].[app_Avatar_GetByID] @AID INT
-AS
-SELECT *
-FROM [Avatar]
-WHERE AID = @AID
