+++ GRA.Database/dbo/Stored Procedures/app_Patron_Insert.sql
@@ -42,9 +42,9 @@ CREATE PROCEDURE [dbo].[app_Patron_Insert] (
 	@Custom3 VARCHAR(50),
 	@Custom4 VARCHAR(50),
 	@Custom5 VARCHAR(50),
-	@AvatarID INT,
 	@SDistrict INT,
 	@DailyGoal INT,
+	@AvatarState VARCHAR(50),
 	@GoalCache INT,
 	@TenID INT = 0,
 	@FldInt1 INT = 0,
@@ -109,9 +109,9 @@ BEGIN
 		Custom3,
 		Custom4,
 		Custom5,
-		AvatarID,
 		SDistrict,
 		DailyGoal,
+		AvatarState,
 		GoalCache,
 		TenID,
 		FldInt1,
@@ -173,9 +173,9 @@ BEGIN
 		@Custom3,
 		@Custom4,
 		@Custom5,
-		@AvatarID,
 		@SDistrict,
 		@DailyGoal,
+		@AvatarState,
 		@GoalCache,
 		@TenID,
 		@FldInt1,
