+++ GRA.Database/dbo/Stored Procedures/app_Patron_Update.sql
@@ -43,9 +43,9 @@ CREATE PROCEDURE [dbo].[app_Patron_Update] (
 	@Custom3 VARCHAR(50),
 	@Custom4 VARCHAR(50),
 	@Custom5 VARCHAR(50),
-	@AvatarID INT,
 	@SDistrict INT,
 	@DailyGoal INT,
+	@AvatarState VARCHAR(50),
 	@GoalCache INT,
 	@TenID INT = 0,
 	@FldInt1 INT = 0,
@@ -108,9 +108,9 @@ SET IsMasterAccount = @IsMasterAccount,
 	Custom3 = @Custom3,
 	Custom4 = @Custom4,
 	Custom5 = @Custom5,
-	AvatarID = @AvatarID,
 	SDistrict = @SDistrict,
 	DailyGoal = @DailyGoal,
+	AvatarState = @AvatarState,
 	GoalCache = @GoalCache,
 	TenID = @TenID,
 	FldInt1 = @FldInt1,
