+++ GRA.Database/dbo/Tables/Patron.sql
@@ -42,11 +42,11 @@
     [Custom3]                  VARCHAR (50)    NULL,
     [Custom4]                  VARCHAR (50)    NULL,
     [Custom5]                  VARCHAR (50)    NULL,
-    [AvatarID]                 INT             NULL,
     [RegistrationDate]         DATETIME        CONSTRAINT [DF_Patron_RegistrationDate] DEFAULT (getdate()) NULL,
     [SDistrict]                INT             NULL,
 	[DailyGoal]                INT             NULL,
 	[GoalCache]                INT             NULL,
+    [AvatarState]              VARCHAR (50)    NULL,
     [TenID]                    INT             NULL,
     [FldInt1]                  INT             NULL,
     [FldInt2]                  INT             NULL,
