+++ SRP/Adventures/Map.aspx.cs
@@ -13,7 +13,7 @@ using GRA.Tools;
 namespace GRA.SRP {
     public partial class GameMap : System.Web.UI.Page {
         private const string StampBasePath = "~/images/Games/Board/stamp_{0}.png";
-        private const string AvatarBasePath = "~/images/Avatars/{0}.png";
+        private const string AvatarBasePath = "~/images/AvatarCache/{0}.png";
         protected void Page_Load(object sender, EventArgs e) {
             var patron = Session[SessionKey.Patron] as Patron;
             if(patron == null) {
@@ -45,7 +45,7 @@ namespace GRA.SRP {
 
             var backImageFile = Server.MapPath(new Logic.Game().GetGameboardPath(patron, gm.PGID));
             var stampImageFile = Server.MapPath(string.Format(StampBasePath, gm.PGID));
-            var avatarImageFile = Server.MapPath(string.Format(AvatarBasePath, patron.AvatarID));
+            var avatarImageFile = Server.MapPath(string.Format(AvatarBasePath, patron.PID));
 
             var newBmp = new System.Drawing.Bitmap(width, height, System.Drawing.Imaging.PixelFormat.Format24bppRgb);
             System.Drawing.Image avatarImage = null;
