+++ SRP/ControlRoom/Controls/PatronCtl.ascx
@@ -86,19 +86,6 @@
             </td>
         <tr/>
 
-
-        <tr>
-            <td>Avatar:</td>
-            <td>
-                <asp:TextBox ID="AvatarIDTxt" runat="server" Text='<%# Eval("AvatarID") %>' Visible="False"></asp:TextBox>
-                <asp:DropDownList ID="AvatarID" runat="server" DataSourceID="odsDDAvatars" DataTextField="Name" DataValueField="AID" 
-                    AppendDataBoundItems="True" 
-                    >
-                    <asp:ListItem Value="0" Text="[Select a Value]"></asp:ListItem>
-                </asp:DropDownList>
-            </td>
-        </tr>
-
         <tr>
             <td>Is Master Account?:</td>
             <td>
