--- SRP_DAL/Avatar.cs
@@ -1,521 +1,0 @@
-﻿using System;
-using System.Data;
-using System.Data.SqlClient;
-using System.Configuration;
-using System.IO;
-using System.Web;
-using System.Web.Security;
-using Microsoft.ApplicationBlocks.Data;
-using System.Collections;
-using GRA.SRP.Core.Utilities;
-using System.Collections.Generic;
-using Newtonsoft.Json;
-
-namespace GRA.SRP.DAL
-{
-
-[Serializable]    public class Avatar : EntityBase
-    {
-        public static new string Version { get { return "2.0"; } }
-
-        #region Private Variables
-
-        private static string conn = GRA.SRP.Core.Utilities.GlobalUtilities.SRPDB;
-
-        private int myAID;
-        private string myName = "";
-        private string myGender = "";
-        private DateTime myLastModDate;
-        private string myLastModUser;
-        private DateTime myAddedDate;
-        private string myAddedUser;
-
-        private int myTenID = 0;
-        private int myFldInt1 = 0;
-        private int myFldInt2 = 0;
-        private int myFldInt3 = 0;
-        private bool myFldBit1 = false;
-        private bool myFldBit2 = false;
-        private bool myFldBit3 = false;
-        private string myFldText1 = "";
-        private string myFldText2 = "";
-        private string myFldText3 = "";
-
-        #endregion
-
-        #region Accessors
-
-        public int AID
-        {
-            get { return myAID; }
-            set { myAID = value; }
-        }
-        public string Name
-        {
-            get { return myName; }
-            set { myName = value; }
-        }
-        public string Gender
-        {
-            get { return myGender; }
-            set { myGender = value; }
-        }
-        public DateTime LastModDate
-        {
-            get { return myLastModDate; }
-            set { myLastModDate = value; }
-        }
-        public string LastModUser
-        {
-            get { return myLastModUser; }
-            set { myLastModUser = value; }
-        }
-        public DateTime AddedDate
-        {
-            get { return myAddedDate; }
-            set { myAddedDate = value; }
-        }
-        public string AddedUser
-        {
-            get { return myAddedUser; }
-            set { myAddedUser = value; }
-        }
-
-        public int TenID
-        {
-            get { return myTenID; }
-            set { myTenID = value; }
-        }
-
-        public int FldInt1
-        {
-            get { return myFldInt1; }
-            set { myFldInt1 = value; }
-        }
-
-        public int FldInt2
-        {
-            get { return myFldInt2; }
-            set { myFldInt2 = value; }
-        }
-
-        public int FldInt3
-        {
-            get { return myFldInt3; }
-            set { myFldInt3 = value; }
-        }
-
-        public bool FldBit1
-        {
-            get { return myFldBit1; }
-            set { myFldBit1 = value; }
-        }
-
-        public bool FldBit2
-        {
-            get { return myFldBit2; }
-            set { myFldBit2 = value; }
-        }
-
-        public bool FldBit3
-        {
-            get { return myFldBit3; }
-            set { myFldBit3 = value; }
-        }
-
-        public string FldText1
-        {
-            get { return myFldText1; }
-            set { myFldText1 = value; }
-        }
-
-        public string FldText2
-        {
-            get { return myFldText2; }
-            set { myFldText2 = value; }
-        }
-
-        public string FldText3
-        {
-            get { return myFldText3; }
-            set { myFldText3 = value; }
-        }
-        #endregion
-
-
-        #region Constructors
-
-        public Avatar()
-        {
-            TenID = (HttpContext.Current.Session["TenantID"] == null || HttpContext.Current.Session["TenantID"].ToString() == "" ? -1 : (int)HttpContext.Current.Session["TenantID"]);
-        }
-
-        #endregion
-
-        #region stored procedure wrappers
-
-        public static DataSet GetAll()
-        {
-            return GetAll(true);
-        }
-
-        public static DataSet GetAll(bool forCurrentTenantOnly = true)
-        {
-            var arrParams = new SqlParameter[1];
-            if (forCurrentTenantOnly)
-            {
-                arrParams[0] = new SqlParameter("@TenID", 
-                    (HttpContext.Current.Session["TenantID"] == null || HttpContext.Current.Session["TenantID"].ToString() == "" ? 
-                            -1 : 
-                            (int) HttpContext.Current.Session["TenantID"])
-                );
-            }
-            else
-            {
-                arrParams[0] = new SqlParameter("@TenID", DBNull.Value);
-            }
-
-            return SqlHelper.ExecuteDataset(conn, CommandType.StoredProcedure, "app_Avatar_GetAll", arrParams);
-        }
-
-        public static DataSet GetAll(int TenID)
-        {
-            var arrParams = new SqlParameter[1];
-            arrParams[0] = new SqlParameter("@TenID", TenID);
-
-            return SqlHelper.ExecuteDataset(conn, CommandType.StoredProcedure, "app_Avatar_GetAll", arrParams);
-        }
-
-        public class DdslickData {
-            public string text { get; set; }
-            public int value { get; set; }
-            public bool selected { get; set; }
-            public string imageSrc { get; set; }
-            public string description { get; set; }
-        }
-
-        public static string  GetJSONForSelection(int selected)
-        {
-            var ds = GetAll();
-            var avatars = new List<DdslickData>();
-            foreach(DataRow row in ds.Tables[0].Rows) {
-                var avatar = new DdslickData {
-                    text = (string)row["Name"],
-                    value = (int)row["AID"],
-                    selected = (int)row["AID"] == selected,
-                    imageSrc = string.Format("/images/Avatars/sm_{0}.png", row["AID"]),
-                    description = "&nbsp;"
-                };
-                avatars.Add(avatar);
-            }
-            return JsonConvert.SerializeObject(avatars.ToArray());
-        }
-
-        public static Avatar FetchObject(int AID)
-        {
-
-            // declare reader
-
-            SqlDataReader dr;
-
-            SqlParameter[] arrParams = new SqlParameter[1];
-
-            arrParams[0] = new SqlParameter("@AID", AID);
-
-            dr = SqlHelper.ExecuteReader(conn, CommandType.StoredProcedure, "app_Avatar_GetByID", arrParams);
-
-            if (dr.Read())
-            {
-
-                // declare return value
-
-                Avatar result = new Avatar();
-
-                DateTime _datetime;
-
-                int _int;
-
-                //decimal _decimal;
-
-                if (int.TryParse(dr["AID"].ToString(), out _int)) result.AID = _int;
-                result.Name = dr["Name"].ToString();
-                result.Gender = dr["Gender"].ToString();
-                if (DateTime.TryParse(dr["LastModDate"].ToString(), out _datetime)) result.LastModDate = _datetime;
-                result.LastModUser = dr["LastModUser"].ToString();
-                if (DateTime.TryParse(dr["AddedDate"].ToString(), out _datetime)) result.AddedDate = _datetime;
-                result.AddedUser = dr["AddedUser"].ToString();
-
-                if (int.TryParse(dr["TenID"].ToString(), out _int)) result.TenID = _int;
-                if (int.TryParse(dr["FldInt1"].ToString(), out _int)) result.FldInt1 = _int;
-                if (int.TryParse(dr["FldInt2"].ToString(), out _int)) result.FldInt2 = _int;
-                if (int.TryParse(dr["FldInt3"].ToString(), out _int)) result.FldInt3 = _int;
-                result.FldBit1 = bool.Parse(dr["FldBit1"].ToString());
-                result.FldBit2 = bool.Parse(dr["FldBit2"].ToString());
-                result.FldBit3 = bool.Parse(dr["FldBit3"].ToString());
-                result.FldText1 = dr["FldText1"].ToString();
-                result.FldText2 = dr["FldText2"].ToString();
-                result.FldText3 = dr["FldText3"].ToString();
-
-                dr.Close();
-
-                return result;
-
-            }
-
-            dr.Close();
-
-            return null;
-
-        }
-
-        public Avatar GetAvatar(int AID)
-        {
-
-            // declare reader
-
-            SqlDataReader dr;
-
-            SqlParameter[] arrParams = new SqlParameter[1];
-
-            arrParams[0] = new SqlParameter("@AID", AID);
-
-            dr = SqlHelper.ExecuteReader(conn, CommandType.StoredProcedure, "app_Avatar_GetByID", arrParams);
-
-            if (dr.Read())
-            {
-
-                // declare return value
-
-                Avatar result = new Avatar();
-
-                DateTime _datetime;
-
-                int _int;
-
-                //decimal _decimal;
-
-                if (int.TryParse(dr["AID"].ToString(), out _int)) result.AID = _int;
-                result.Name = dr["Name"].ToString();
-                result.Gender = dr["Gender"].ToString();
-                if (DateTime.TryParse(dr["LastModDate"].ToString(), out _datetime)) result.LastModDate = _datetime;
-                result.LastModUser = dr["LastModUser"].ToString();
-                if (DateTime.TryParse(dr["AddedDate"].ToString(), out _datetime)) result.AddedDate = _datetime;
-                result.AddedUser = dr["AddedUser"].ToString();
-
-                if (int.TryParse(dr["TenID"].ToString(), out _int)) result.TenID = _int;
-                if (int.TryParse(dr["FldInt1"].ToString(), out _int)) result.FldInt1 = _int;
-                if (int.TryParse(dr["FldInt2"].ToString(), out _int)) result.FldInt2 = _int;
-                if (int.TryParse(dr["FldInt3"].ToString(), out _int)) result.FldInt3 = _int;
-                result.FldBit1 = bool.Parse(dr["FldBit1"].ToString());
-                result.FldBit2 = bool.Parse(dr["FldBit2"].ToString());
-                result.FldBit3 = bool.Parse(dr["FldBit3"].ToString());
-                result.FldText1 = dr["FldText1"].ToString();
-                result.FldText2 = dr["FldText2"].ToString();
-                result.FldText3 = dr["FldText3"].ToString();
-
-                dr.Close();
-
-                return result;
-
-            }
-
-            dr.Close();
-
-            return null;
-
-        }
-
-        public bool Fetch(int AID)
-        {
-
-            // declare reader
-
-            SqlDataReader dr;
-
-            SqlParameter[] arrParams = new SqlParameter[1];
-
-            arrParams[0] = new SqlParameter("@AID", AID);
-
-            dr = SqlHelper.ExecuteReader(conn, CommandType.StoredProcedure, "app_Avatar_GetByID", arrParams);
-
-            if (dr.Read())
-            {
-
-                // declare return value
-
-                DateTime _datetime;
-
-                int _int;
-
-                //decimal _decimal;
-
-                if (int.TryParse(dr["AID"].ToString(), out _int)) this.AID = _int;
-                this.Name = dr["Name"].ToString();
-                this.Gender = dr["Gender"].ToString();
-                if (DateTime.TryParse(dr["LastModDate"].ToString(), out _datetime)) this.LastModDate = _datetime;
-                this.LastModUser = dr["LastModUser"].ToString();
-                if (DateTime.TryParse(dr["AddedDate"].ToString(), out _datetime)) this.AddedDate = _datetime;
-                this.AddedUser = dr["AddedUser"].ToString();
-
-                if (int.TryParse(dr["TenID"].ToString(), out _int)) this.TenID = _int;
-                if (int.TryParse(dr["FldInt1"].ToString(), out _int)) this.FldInt1 = _int;
-                if (int.TryParse(dr["FldInt2"].ToString(), out _int)) this.FldInt2 = _int;
-                if (int.TryParse(dr["FldInt3"].ToString(), out _int)) this.FldInt3 = _int;
-                this.FldBit1 = bool.Parse(dr["FldBit1"].ToString());
-                this.FldBit2 = bool.Parse(dr["FldBit2"].ToString());
-                this.FldBit3 = bool.Parse(dr["FldBit3"].ToString());
-                this.FldText1 = dr["FldText1"].ToString();
-                this.FldText2 = dr["FldText2"].ToString();
-                this.FldText3 = dr["FldText3"].ToString();
-
-                dr.Close();
-
-                return true;
-            }
-
-            dr.Close();
-
-            return false;
-
-        }
-
-
-
-        public int Insert()
-        {
-
-            return Insert(this);
-
-        }
-
-        public static int Insert(Avatar o)
-        {
-
-            SqlParameter[] arrParams = new SqlParameter[17];
-
-            arrParams[0] = new SqlParameter("@Name", GRA.SRP.Core.Utilities.GlobalUtilities.DBSafeValue(o.Name, o.Name.GetTypeCode()));
-            arrParams[1] = new SqlParameter("@Gender", GRA.SRP.Core.Utilities.GlobalUtilities.DBSafeValue(o.Gender, o.Gender.GetTypeCode()));
-            arrParams[2] = new SqlParameter("@LastModDate", GRA.SRP.Core.Utilities.GlobalUtilities.DBSafeValue(o.LastModDate, o.LastModDate.GetTypeCode()));
-            arrParams[3] = new SqlParameter("@LastModUser", GRA.SRP.Core.Utilities.GlobalUtilities.DBSafeValue(o.LastModUser, o.LastModUser.GetTypeCode()));
-            arrParams[4] = new SqlParameter("@AddedDate", GRA.SRP.Core.Utilities.GlobalUtilities.DBSafeValue(o.AddedDate, o.AddedDate.GetTypeCode()));
-            arrParams[5] = new SqlParameter("@AddedUser", GRA.SRP.Core.Utilities.GlobalUtilities.DBSafeValue(o.AddedUser, o.AddedUser.GetTypeCode()));
-
-            arrParams[6] = new SqlParameter("@TenID", GRA.SRP.Core.Utilities.GlobalUtilities.DBSafeValue(o.TenID, o.TenID.GetTypeCode()));
-            arrParams[7] = new SqlParameter("@FldInt1", GRA.SRP.Core.Utilities.GlobalUtilities.DBSafeValue(o.FldInt1, o.FldInt1.GetTypeCode()));
-            arrParams[8] = new SqlParameter("@FldInt2", GRA.SRP.Core.Utilities.GlobalUtilities.DBSafeValue(o.FldInt2, o.FldInt2.GetTypeCode()));
-            arrParams[9] = new SqlParameter("@FldInt3", GRA.SRP.Core.Utilities.GlobalUtilities.DBSafeValue(o.FldInt3, o.FldInt3.GetTypeCode()));
-            arrParams[10] = new SqlParameter("@FldBit1", GRA.SRP.Core.Utilities.GlobalUtilities.DBSafeValue(o.FldBit1, o.FldBit1.GetTypeCode()));
-            arrParams[11] = new SqlParameter("@FldBit2", GRA.SRP.Core.Utilities.GlobalUtilities.DBSafeValue(o.FldBit2, o.FldBit2.GetTypeCode()));
-            arrParams[12] = new SqlParameter("@FldBit3", GRA.SRP.Core.Utilities.GlobalUtilities.DBSafeValue(o.FldBit3, o.FldBit3.GetTypeCode()));
-            arrParams[13] = new SqlParameter("@FldText1", GRA.SRP.Core.Utilities.GlobalUtilities.DBSafeValue(o.FldText1, o.FldText1.GetTypeCode()));
-            arrParams[14] = new SqlParameter("@FldText2", GRA.SRP.Core.Utilities.GlobalUtilities.DBSafeValue(o.FldText2, o.FldText2.GetTypeCode()));
-            arrParams[15] = new SqlParameter("@FldText3", GRA.SRP.Core.Utilities.GlobalUtilities.DBSafeValue(o.FldText3, o.FldText3.GetTypeCode()));
-
-            arrParams[16] = new SqlParameter("@AID", GRA.SRP.Core.Utilities.GlobalUtilities.DBSafeValue(o.AID, o.AID.GetTypeCode()));
-            arrParams[16].Direction = ParameterDirection.Output;
-
-            SqlHelper.ExecuteNonQuery(conn, CommandType.StoredProcedure, "app_Avatar_Insert", arrParams);
-            o.AID = int.Parse(arrParams[16].Value.ToString());
-            return o.AID;
-
-        }
-
-        public int Update()
-        {
-
-            return Update(this);
-
-        }
-
-        public static int Update(Avatar o)
-        {
-
-            int iReturn = -1; //assume the worst
-
-            SqlParameter[] arrParams = new SqlParameter[17];
-
-            arrParams[0] = new SqlParameter("@AID", GRA.SRP.Core.Utilities.GlobalUtilities.DBSafeValue(o.AID, o.AID.GetTypeCode()));
-            arrParams[1] = new SqlParameter("@Name", GRA.SRP.Core.Utilities.GlobalUtilities.DBSafeValue(o.Name, o.Name.GetTypeCode()));
-            arrParams[2] = new SqlParameter("@Gender", GRA.SRP.Core.Utilities.GlobalUtilities.DBSafeValue(o.Gender, o.Gender.GetTypeCode()));
-            arrParams[3] = new SqlParameter("@LastModDate", GRA.SRP.Core.Utilities.GlobalUtilities.DBSafeValue(o.LastModDate, o.LastModDate.GetTypeCode()));
-            arrParams[4] = new SqlParameter("@LastModUser", GRA.SRP.Core.Utilities.GlobalUtilities.DBSafeValue(o.LastModUser, o.LastModUser.GetTypeCode()));
-            arrParams[5] = new SqlParameter("@AddedDate", GRA.SRP.Core.Utilities.GlobalUtilities.DBSafeValue(o.AddedDate, o.AddedDate.GetTypeCode()));
-            arrParams[6] = new SqlParameter("@AddedUser", GRA.SRP.Core.Utilities.GlobalUtilities.DBSafeValue(o.AddedUser, o.AddedUser.GetTypeCode()));
-            arrParams[7] = new SqlParameter("@TenID", GRA.SRP.Core.Utilities.GlobalUtilities.DBSafeValue(o.TenID, o.TenID.GetTypeCode()));
-            arrParams[8] = new SqlParameter("@FldInt1", GRA.SRP.Core.Utilities.GlobalUtilities.DBSafeValue(o.FldInt1, o.FldInt1.GetTypeCode()));
-            arrParams[9] = new SqlParameter("@FldInt2", GRA.SRP.Core.Utilities.GlobalUtilities.DBSafeValue(o.FldInt2, o.FldInt2.GetTypeCode()));
-            arrParams[10] = new SqlParameter("@FldInt3", GRA.SRP.Core.Utilities.GlobalUtilities.DBSafeValue(o.FldInt3, o.FldInt3.GetTypeCode()));
-            arrParams[11] = new SqlParameter("@FldBit1", GRA.SRP.Core.Utilities.GlobalUtilities.DBSafeValue(o.FldBit1, o.FldBit1.GetTypeCode()));
-            arrParams[12] = new SqlParameter("@FldBit2", GRA.SRP.Core.Utilities.GlobalUtilities.DBSafeValue(o.FldBit2, o.FldBit2.GetTypeCode()));
-            arrParams[13] = new SqlParameter("@FldBit3", GRA.SRP.Core.Utilities.GlobalUtilities.DBSafeValue(o.FldBit3, o.FldBit3.GetTypeCode()));
-            arrParams[14] = new SqlParameter("@FldText1", GRA.SRP.Core.Utilities.GlobalUtilities.DBSafeValue(o.FldText1, o.FldText1.GetTypeCode()));
-            arrParams[15] = new SqlParameter("@FldText2", GRA.SRP.Core.Utilities.GlobalUtilities.DBSafeValue(o.FldText2, o.FldText2.GetTypeCode()));
-            arrParams[16] = new SqlParameter("@FldText3", GRA.SRP.Core.Utilities.GlobalUtilities.DBSafeValue(o.FldText3, o.FldText3.GetTypeCode()));
-
-            try
-            {
-
-                iReturn = SqlHelper.ExecuteNonQuery(conn, CommandType.StoredProcedure, "app_Avatar_Update", arrParams);
-
-            }
-
-            catch (SqlException exx)
-            {
-
-                System.Diagnostics.Debug.Write(exx.Message);
-
-            }
-
-            return iReturn;
-
-        }
-
-        public int Delete()
-        {
-
-            return Delete(this);
-
-        }
-
-        public static int Delete(Avatar o)
-        {
-
-            int iReturn = -1; //assume the worst
-
-            var arrParams = new SqlParameter[2];
-            arrParams[0] = new SqlParameter("@AID", GRA.SRP.Core.Utilities.GlobalUtilities.DBSafeValue(o.AID, o.AID.GetTypeCode()));
-            arrParams[1] = new SqlParameter("@TenID",
-                (HttpContext.Current.Session["TenantID"] == null || HttpContext.Current.Session["TenantID"].ToString() == "" ?
-                        -1 :
-                        (int)HttpContext.Current.Session["TenantID"])
-            );
-
-            try
-            {
-                var fileName = (HttpContext.Current.Server.MapPath("~/Images/Avatars/") + "\\" + o.AID.ToString() + ".png");
-                File.Delete(fileName);
-                fileName = (HttpContext.Current.Server.MapPath("~/Images/Avatars/") + "\\sm_" + o.AID.ToString() + ".png"); 
-                File.Delete(fileName);
-                fileName = (HttpContext.Current.Server.MapPath("~/Images/Avatars/") + "\\md_" + o.AID.ToString() + ".png"); 
-                File.Delete(fileName);
-
-                iReturn = SqlHelper.ExecuteNonQuery(conn, CommandType.StoredProcedure, "app_Avatar_Delete", arrParams);
-            }
-
-            catch (SqlException exx)
-            {
-
-                System.Diagnostics.Debug.Write(exx.Message);
-
-            }
-
-            return iReturn;
-
-        }
-
-        #endregion
-
-    }//end class
-
-}//end namespace
-
