+++ SRP_DAL/SRP_DAL.csproj
@@ -49,7 +49,6 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Avatar.cs" />
     <Compile Include="AvatarPart.cs" />
     <Compile Include="Award.cs" />
     <Compile Include="Badge.cs" />
